Fixed issues with missing space in circle of animals

+++ core/src/com/github/samelvhatargh/druid/systems/CollectedAnimalsPositionCalculator.kt
@@ -4,7 +4,6 @@ import com.badlogic.ashley.core.Entity
 import com.badlogic.ashley.systems.SortedIteratingSystem
 import com.github.samelvhatargh.druid.components.Animal
 import com.github.samelvhatargh.druid.components.CollectedAnimal
-import com.github.samelvhatargh.druid.components.Druid
 import com.github.samelvhatargh.druid.components.Position
 import com.github.samelvhatargh.druid.getDruid
 import ktx.ashley.allOf
@@ -28,6 +27,6 @@ class CollectedAnimalsPositionCalculator :
         val collectedAnimal = entity[CollectedAnimal.mapper]!!
 
         position.vec.set(0f, druid.radius)
-        position.vec.rotateDeg(collectedAnimal.order * (360 / druid.animalsCount).toFloat() + druid.angle)
+        position.vec.rotateDeg(collectedAnimal.order * (360f / druid.animalsCount) + druid.angle)
     }
 }
